rename data.js -> src/data.js. #19
